bsd: refactor common netbsd and freebsd method up in BSD distro class
There is a lot of duplication of code that could be shared between
the FreeBSD and NetBSD classes. Let's avoid as much boilerplate
cut-n-paste as possible.

Move the following methods up into the parent BSD class:
 * create_group which now uses group_add_cmd_prefix and
   _get_add_member_to_group_cmd to specialize subclass commands
 * install_packages can be generalized since NetBSD.update_package_sources
   is nothing
 * package_command is generalized with minor command diffs sourced from
   class attributes pkg_cmd_install_prefix and pkg_cmd_remove_prefix and
   custom environment given by _get_pkg_cmd_environ

Also in this commit:
 * A bug fix in existing Freebsd which was switching group name and
   and username parameters
 * I left a TODO message for your review as I don't think either BSD class
 handles a distro.package_command('upgrade') call and it might be needed
 in a followup PR
